refactor role (#2)
