+++ src/user/controllers/role_controller.py
@@ -1,16 +1,16 @@
-from fastapi import APIRouter, Depends
-from sqlalchemy.orm import Session
+from fastapi import APIRouter
 
-from src.config.database.db_helper import db_helper
-from src.user.servises.role_servise import RoleService
-from src.user.schema.role_schema import CreateRole
+from ..dependencies.services import IRoleService
+from src.user.dtos.role_dto import CreateRole
 
 router = APIRouter(prefix="/role", tags=["role"])
 
 
 @router.post("/")
 async def create_role(
-        data: CreateRole,
-        db_session: Session = Depends(db_helper.get_session),
+    dto: CreateRole,
+    service: IRoleService
 ):
-    return await RoleService().create(data, db_session)
+    return await service.create(dto)
+
+
