+++ src/user/dependencies/services.py
@@ -1,6 +1,9 @@
 from typing import Annotated
 from fastapi import Depends
-from ..services.user_servise import UserService
+
+from ..services.role_service import RoleService
+from ..services.user_service import UserService
 
 
 IUserService = Annotated[UserService, Depends()]
+IRoleService = Annotated[RoleService, Depends()]
