+++ src/user/services/user_service.py
@@ -10,7 +10,6 @@ class UserService:
 
     async def create(self, dto: CreateUser):
         user = UserEntity(**dto.model_dump()).get_new_hash_password()
-        print("$$"*100, user)
         return await self.repository.create(user)
 
     async def get(self, pk: int):
